requirements.txt
streamlit==1.35.0
pandas==2.2.2
scikit-learn==1.4.1.post1
python-dotenv==1.0.1
requests==2.31.0
